requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autogen-agentchat==0.6.4
autogen-core==0.6.4
autogen-ext==0.6.4
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b10
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-search-documents==11.5.2
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.15
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
cssselect==1.3.0
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
email_validator==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
hubspot-api-client==11.1.0
huggingface-hub==0.31.1
humanfriendly==10.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
json-schema-to-pydantic==0.2.4
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
litellm==1.68.1
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.74.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.7
playwright==1.52.0
posthog==4.0.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytz==2025.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rank-bm25==0.2.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==5.0.0
setuptools==75.8.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
sse-starlette==2.2.1
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.53.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
